# openapi.yaml
openapi: 3.0.3
info:
  title: MikroTik Dashboard API
  description: REST API for monitoring MikroTik RouterOS devices
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080
    description: Local dashboard server
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie obtained from /api/login
  schemas:
    SystemInfo:
      type: object
      properties:
        identity: { type: string }
        version: { type: string }
        board: { type: string }
        uptime: { type: string }
        cpu_load: { type: integer }
        free_memory: { type: integer }
        total_memory: { type: integer }
        free_hdd: { type: integer }
        total_hdd: { type: integer }
        temperature: { type: integer, nullable: true }
...
    Error:
      type: object
      properties:
        error: { type: string }
security:
  - CookieAuth: []
paths:
  /api/login:
    post:
      summary: Authenticate with dashboard token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        200:
          description: Authentication successful, session cookie set
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        403:
          description: Invalid token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/system:
    get:
      summary: Get system information
      responses:
        200:
          description: System info
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SystemInfo' }
        401:
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...